==> Code/00_data_cleaning_merging.ipynb <==
## Load packages

import pandas as pd
import numpy as np
import random
import re
import os

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

from scipy.stats import zscore



## Load and inspect data 

sentencing_data_cleaned = pd.read_csv("../Data/sentencing_cleaned.csv", low_memory = False)

sentencing_data_cleaned.head()
sentencing_data_cleaned.shape
sentencing_data_cleaned.info

## Create Term Length and Sentence Length Variables

## Creating Term Length Variable

## subsetting to just units of days, years, months, and life sentences (Note: we are equating one life sentence to be 100 years)

sentencing_df = sentencing_data_cleaned[sentencing_data_cleaned["COMMITMENT_UNIT"].isin(["Year(s)", "Natural Life", "Days", "Months"]) & sentencing_data_cleaned["COMMITMENT_TYPE"].isin(["Illinois Department of Corrections", "Cook County Department of Corrections"])]


sentencing_df["COMMITMENT_TYPE"].unique()
sentencing_df[["COMMITMENT_UNIT", "COMMITMENT_TERM"]][sentencing_df["COMMITMENT_TYPE"] == "Cook County Boot Camp"]




## Creating sentence length variable (in years)

sentencing_df["COMMITMENT_TERM"] = sentencing_df["COMMITMENT_TERM"].astype(float)


sentencing_df["sentence_length"] = sentencing_df.apply(lambda row: row["COMMITMENT_TERM"] if row["COMMITMENT_UNIT"] == "Year(s)" else  
                                                       row["COMMITMENT_TERM"] * 100 if row["COMMITMENT_UNIT"] == "Natural Life" else
                                                       row["COMMITMENT_TERM"] / 365 if row["COMMITMENT_UNIT"] == "Days" else 
                                                       row["COMMITMENT_TERM"] / 12 if row["COMMITMENT_UNIT"] == "Months" else None, axis = 1)


## Make sentencing unit standardizable by creating Z score 

## Need to convert sentencing unit to something standardizable across crimes 
## To do this will use a z score (how many standard deviations something is from the mean) grouped by the exact same charge type


sentencing_df["sentence_length_zscore"] = sentencing_df.groupby("DISPOSITION_CHARGED_OFFENSE_TITLE")["sentence_length"].transform(zscore)

## looking at z score stats just quickly

np.mean(sentencing_df["sentence_length_zscore"])

## 0 z score would mean perfectly average or neutral sentence length for the crime


## Export data as csv file to Data folder 

sentencing_df.to_csv(r'../Data/sentencing_data_withZ.csv')
