==> Code/00_inspecting_data.ipynb <==
## Investigating Initial Dataset

## Load packages

import pandas as pd
import numpy as np
import random
import re
import os

## Note the code below is not a function, it just resets the output so it shows all lines


from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"


## Loading data 

sentencing_data_cleaned = pd.read_csv("../Data/sentencing_cleaned.csv", low_memory = False)

## Initial Inspecting Data

print(sentencing_data_cleaned.head())
print(sentencing_data_cleaned.shape)
print(sentencing_data_cleaned.info)


## Simplify Data For Summary Table

## Summary table subset to include variables relevant to our analysis + variables that show number of observations in the dataset as a whole 

table_df = sentencing_data_cleaned[["CASE_ID", "sentenceymd_derived", "age_derived", "CASE_PARTICIPANT_ID", "DISPOSITION_CHARGED_OFFENSE_TITLE", "simplified_offense_derived", "judgeid_derived", "RACE", "CHARGE_DISPOSITION", "GENDER", "COMMITMENT_TYPE", "COMMITMENT_UNIT", "COMMITMENT_TERM", "PRIMARY_CHARGE_FLAG"]]

## Converting Confusing Numerics to Strings:

table_df[["CASE_ID", "CASE_PARTICIPANT_ID"]] = str(["CASE_ID", "CASE_PARTICIPANT_ID"])

## Converting Dates (sentenceymd_derived) to Datetime:

table_df["sentenceymd_derived"] = pd.to_datetime(table_df["sentenceymd_derived"])

## Create Simplified Summary Table for Data

## Creating simplified summary table for data: 

summary_table = table_df.describe(include = ["datetime", "int", "float", "object"], datetime_is_numeric = True)

## Verify Table Was Created Properly 

print(summary_table)


## Exporting Summary Table 

summary_table.to_csv(r'../Output/Tables/Simplified_Summary_Table.csv')
