==> Code/02_creating_regression_variables.ipynb <==
## Creating Additional Columns (Independent Variables) Needed For Analysis

## Load packages

import pandas as pd
import numpy as np
import random
import re
import os
import matplotlib.pyplot as plt

## Note the code below is not a function, it just resets the output so it shows all lines

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## Load and Inspect Cleaned Data (with Z Scores)

sentencing_df = pd.read_csv("../Data/sentencing_data_withZ.csv", low_memory = False)

print(sentencing_df.head(), sentencing_df.info(), sentencing_df.shape)

## Changing existing boolean race and gender data to binary

## for regression, is_(race)_derived and is_male_derived variables need to be binary

sentencing_df[["is_black_derived", "is_white_derived", "is_hisp_derived", "is_other_derived", "is_male_derived"]] = sentencing_df[["is_black_derived", "is_white_derived", "is_hisp_derived", "is_other_derived", "is_male_derived"]].astype(int)


## Checking to make sure binary variables created properly

print(sentencing_df[["is_black_derived", "is_other_derived", "is_hisp_derived", "is_other_derived", "is_male_derived"]])

## Creating is_guilty_plea and is_innocent_plea

## Need to create a binary variable for innocent and guilty pleas out of CHARGE_DISPOSITION

## Binary Guilty: 

sentencing_df["is_guilty_plea"] = np.where(sentencing_df["CHARGE_DISPOSITION"] == "Plea Of Guilty", 1, 0)

## Binary Not Guilty:

sentencing_df["is_innocent_plea"] = np.where(sentencing_df["CHARGE_DISPOSITION"] == "Plea Of Guilty", 0, 1)


## Checking to make sure the variable worked: 

print(sentencing_df[["is_guilty_plea", "is_innocent_plea", "CHARGE_DISPOSITION"]])

## Creating is_female_derived (is_male_derived already exists)

## Need to create a binary variable for females to complement is_male

sentencing_df["is_female_derived"] = np.where(sentencing_df["GENDER"] == "Female", 1, 0)

## Checking to make sure it worked:

print(sentencing_df[["is_female_derived", "GENDER"]])


## Creating Nth Case (To measure how judicial experience corresponds to severity) 

## Creating a variable for each case that indicates how many cases the assigned judge has seen (measure of judicial experience)
## Did not make this into a function because the task is too specific

## Casting dates to datetime

sentencing_df["sentenceymd_derived_dt"] = pd.to_datetime(sentencing_df["sentenceymd_derived"])

## Sorting Values by Sentence Date and Judge: 

sentencing_df_sorted = sentencing_df.sort_values(["SENTENCE_JUDGE", "sentenceymd_derived_dt"])

## checking to make suresorting worked 

print(sentencing_df_sorted.head)

## Creating column for Nth case (nth_case)

sentencing_df_sorted["nth_case"] = sentencing_df_sorted.groupby("SENTENCE_JUDGE").cumcount() + 1

## checking to make sure nth case works:

print(sentencing_df_sorted[["nth_case", "SENTENCE_JUDGE", "sentenceymd_derived_dt"]])


## Exporting CSV to Data folder

## Data to be used in regression and visualizations

sentencing_df_sorted.to_csv(r'../Data/sentencing_data_for_analysis.csv')
