==> Code/03_regression.ipynb <==
## Load Packages: 

import pandas as pd
import numpy as np
import random
import re
import os
import matplotlib.pyplot as plt

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import statsmodels.api as sm
from statsmodels.iolib.summary2 import summary_col


## Load in and Inspect Regression Data:

sentencing_data_cleaned = pd.read_csv("../Data/sentencing_data_for_regression.csv", low_memory = False)

print(sentencing_data_cleaned.head, sentencing_data_cleaned.shape, sentencing_data_cleaned.info)


## Fixing import issue with converting dates to datetime 

sentencing_data_cleaned["sentenceymd_derived_dt"] = pd.to_datetime(sentencing_data_cleaned["sentenceymd_derived_dt"])

## checking to make sure this conversion worked: 

print(type(sentencing_data_cleaned["sentenceymd_derived_dt"][1]))

## Computing Linear Regression of Continuous Variables

## Using variables created previously vs Sentence Severity

## Dropping nas in the rows of the variables we will model 

age_df = sentencing_data_cleaned.dropna(subset = ["age_derived", "is_guilty_plea", "nth_case", "is_female_derived", "is_innocent_plea",
                                                 "is_male_derived", "is_black_derived", "is_white_derived", 
                                                  "is_hisp_derived", "is_other_derived"])


## Creating the model 
x = age_df[["age_derived", "is_guilty_plea", "nth_case", "is_female_derived", "is_innocent_plea", "is_male_derived", "is_black_derived", "is_white_derived", 
                                                  "is_hisp_derived", "is_other_derived"]]

y = age_df["sentence_length_zscore"]

## creating a constant 

x = sm.add_constant(x)

model = sm.OLS(y, x).fit()

## checking model results 

print(model.summary())


## Combining Model Results as a Table:

table = summary_col(model, stars = True, model_names = ["Sentence Severity"])

print(table)

## Exporting Regression table 

table_df = table.tables[0]

table_df.to_csv(r'../Output/Tables/OLS_table.csv')
