==> Code/04_ols_model.ipynb <==
## Modeling Analysis Data To Look for Causality Between Our Descriptive Variables and Sentence Severity (OLS Model)

## Load Packages: 

import pandas as pd
import numpy as np
import random
import re
import os

## For OLS Model: 

import statsmodels.api as sm

## Note the code below is not a function, it just resets the output so it shows all lines

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"



## Functions

## summary_col() creates summary table(s) from multiple regressions 

from statsmodels.iolib.summary2 import summary_col

"""
Statsmodels.iolib.summary2 source docstring for summary_col: 


    Summarize multiple results instances horizontally (side-by-side).

    Parameters
    ----------
    results : list[Result]
        List of results instances to summarize.
    float_format : str, optional
        String formatting code to use for float results.
    model_names : list[str], optional
        List of model names. If not None, then the number of model names
        should equal the number of results instances in the results list.
    stars : bool, optional
        Whether to include significance stars.
    info_dict : dict or None, optional
        Dictionary of additional information to include in the table. The
        dictionary should be structured so that the keys are column names
        and the values are lists or arrays with the same number of elements
        as there are models being summarized.
    regressor_order : list[str] or None, optional
        List of regressor names in the order they should appear in the
        output table. If not provided, then regressors will be displayed
        in the order they appear in the model results summary tables.
    drop_omitted : bool, optional
        Whether to drop columns for omitted variables.
    title : str or None, optional
        Title to include at the top of the table.
    return_txt : bool, optional
        Whether to return the summary table as a plain-text string.

    Returns
    -------
    table : Summary
        A summary table instance.
    """



## ols() creates a basic Linear Model (OLS) for the x and y columns (variables) inputted 

def ols(x_col, y_col, df):
    
    if not isinstance(x_col, list) & isinstance(y_col, list):
        raise TypeError("X and Y variables must be passed as lists")
    

    ## get rid of nas that would interrupt the model 
    df_no_nas = df.dropna(subset = x_col + y_col)


    ## set x and y equal to correct columns in df 
    x = df_no_nas[x_col]
    y = df_no_nas[y_col]

    ## add a constant to model all independent variables at once
    x = sm.add_constant(x)

    ## fit and create model
    model = sm.OLS(y, x).fit()

    return model
    
    """
    Use ordinary least squares regression to estimate the linear relationship between independent (x) and dependent variables (y)
    

    Parameters:
    
    x_col: List
        List containing column names of independent variable(s)
    
    y_col: List
        List containing column name for dependent variable(s)
    
    df: pandas.DataFrame 
        DataFrame containing the dataset to be modeld 

    Outputs: statsmodels.regression.linear_model.RegressionResultsWrapper
        OLS Model 
    
   TypeError 
        If x_col and y_col are not passed as lists.
    
    """


## Load in and Inspect Analysis Data:

sentencing_data_cleaned = pd.read_csv("../Data/sentencing_data_for_analysis.csv", low_memory = False)

print(sentencing_data_cleaned.head, sentencing_data_cleaned.shape, sentencing_data_cleaned.info)



## Computing Linear Regression of Continuous Variables

## Using Descriptive Variables We Identified vs Sentence Severity

## X and Y Variables:
## ols function must be passed lists

x_vars = ["age_derived", "is_guilty_plea", "nth_case", "is_female_derived", "is_innocent_plea", "is_male_derived", "is_black_derived", "is_white_derived", 
                                                  "is_hisp_derived", "is_other_derived"]
y_var = ["sentence_length_zscore"]

## Creating the model 

model = ols(x_vars, y_var, sentencing_data_cleaned)

## checking model results 

print(model.summary2())



## Combining Model Results Into A Simple Table:

## Use summary_col() to create a simplified table that can be exported for paper

table = summary_col(model, stars = True, model_names = ["Sentence Severity"])

## Check table before exporting
print(table)


## Exporting Regression table 

## Exporting Table
table_df = table.tables[0]

table_df.to_csv(r'../Output/Tables/OLS_table.csv')
